Conflicts:
	framework/src/minko/component/PerspectiveCamera.hpp
	premake4.lua

Former-commit-id: a08f18a8ed19ff7f34bcc04ed44ed1c5f895673f

+++ plugins/particles/src/minko/particle/shape/Cone.cpp
@@ -31,7 +31,7 @@ void
 Cone::initParticle(ParticleData& particle,
 				   bool direction) const
 {
-	float theta		= (rand01() * 2 - 1) * PI;
+	float theta		= (rand01() * 2.f - 1.f) * (float)PI;
 	
 	float cosTheta	= cos(theta);
 	float sinTheta	= sin(theta);
