+++ plugins/particles/src/minko/particle/shape/Cylinder.cpp
@@ -16,7 +16,7 @@ Cylinder::Cylinder(float	height,
 void
 Cylinder::initPosition(ParticleData& particle) const
 {
-	float theta		= (rand01() * 2 - 1) * PI;
+	float theta		= (rand01() * 2.f - 1.f) * (float)PI;
 	
 	float cosTheta	= cos(theta);
 	float sinTheta	= sin(theta);
